Refactor detectBackend and move it to the webdriverio package
