+++ examples/wdio/custom-service/mocha.test.js
@@ -3,6 +3,6 @@ describe('webdriver.io page', () => {
 
     it('should have the right title', () => {
         browser.url('https://webdriver.io')
-        expect(browser).toHaveTitle('WebdriverIO · Next-gen browser and mobile automation test framework for Node.js')
+        expect(browser).toHaveTitle('WebdriverIO · Next-gen browser and mobile automation test framework for Node.js | WebdriverIO')
     })
 })
