+++ examples/wdio/mocha/wdio.conf.js
@@ -33,7 +33,7 @@ exports.config = {
 
     mochaOpts: {
         ui: 'bdd',
-        timeout: 5000
+        timeout: 15000
     },
 
     /**
