+++ packages/wdio-config/src/index.ts
@@ -1,13 +1,11 @@
 /* istanbul ignore file */
 
 import ConfigParser from './lib/ConfigParser'
-import { validateConfig, getSauceEndpoint, detectBackend, isCloudCapability, ModuleRequireService } from './utils'
+import { validateConfig, isCloudCapability, ModuleRequireService } from './utils'
 import { DEFAULT_CONFIGS } from './constants'
 
 export {
     validateConfig,
-    getSauceEndpoint,
-    detectBackend,
     isCloudCapability,
     ConfigParser,
