+++ packages/wdio-config/tests/__fixtures__/wdio.conf.multiremote.rdc.ts
@@ -8,7 +8,7 @@ export const config: ConfigOptions = {
     key: '50fa142c-3121-4gb0-9p07-8q326vvbq7b0',
     specs: [path.join(TEST_ROOT, '*.test.ts')],
     exclude: [
-        path.join(TEST_ROOT, '/detectBackend.test.ts'),
+        path.join(TEST_ROOT, '/RequireLibrary.test.ts'),
         path.join(TEST_ROOT, '/validateConfig.test.ts')
     ],
     region: 'eu',
@@ -22,7 +22,7 @@ export const config: ConfigOptions = {
     }],
     suites: {
         unit: [path.join(TEST_ROOT, 'configparser.test.ts')],
-        mobile: [path.join(TEST_ROOT, 'detectBackend.test.ts')],
+        mobile: [path.join(TEST_ROOT, 'RequireLibrary.test.ts')],
         functional: [
             path.join(TEST_ROOT, 'validateConfig.test.ts'),
             path.join(TEST_ROOT, '..', 'src/index.ts')
