+++ packages/wdio-config/tests/__mocks__/@wdio/config.ts
@@ -1,9 +1,5 @@
 import { DEFAULT_CONFIGS as DEFAULT_CONFIGS_IMPORT } from '../../../src/constants'
-import {
-    getSauceEndpoint as getSauceEndpointMock,
-    isCloudCapability as isCloudCapabilityMock,
-    detectBackend as detectBackendMock
-} from '../../../src/utils'
+import { isCloudCapability as isCloudCapabilityMock } from '../../../src/utils'
 
 class ConfigParserMock {
     addService = jest.fn()
@@ -30,9 +26,7 @@ class ConfigParserMock {
 
 export const ConfigParser = ConfigParserMock
 export const DEFAULT_CONFIGS = DEFAULT_CONFIGS_IMPORT
-export const getSauceEndpoint = jest.fn().mockImplementation(getSauceEndpointMock)
 export const isCloudCapability = jest.fn().mockImplementation(isCloudCapabilityMock)
-export const detectBackend = jest.fn().mockImplementation(detectBackendMock)
 export const validateConfig = jest.fn().mockImplementation(
     (_, config) => Object.assign(
         DEFAULT_CONFIGS_IMPORT(),
