+++ packages/webdriverio/tests/utils/detectBackend.test.ts
@@ -1,4 +1,4 @@
-import { detectBackend } from '../src/utils'
+import detectBackend from '../../src/utils/detectBackend'
 
 describe('detectBackend', () => {
     it('should not set anything if host is set in caps', () => {
